+++ Runtimes/HtmlRun.Terminal/Instructions/SetCursorPositionCmd.cs
@@ -1,0 +1,17 @@
+using HtmlRun.Runtime.Interfaces;
+using HtmlRun.Runtime.Native;
+
+namespace HtmlRun.Terminal.Instructions;
+
+class SetCursorPosition : INativeInstruction
+{
+  public string Key => "SetCursorPosition";
+
+  public Action<ICurrentInstructionContext> Action
+  {
+    get
+    {
+      return ctx => Console.SetCursorPosition(ctx.GetRequiredArgument<int>(0), ctx.GetRequiredArgument<int>(1));
+    }
+  }
+}